fix selectItem signature
Finally found the address for the select item hook. The select item
call (which looks to be a targeting routine with some housing
functinality) happens in a different place than where it did in 6.X and
a lot of additional functionality was implemented into the call hence
why the signature took so long to locate.

This commit does make the plugin work now but there does still seem to
be an issue with the acutal placement. While the items do have their
position set, the placement does not seem to stick. I believe this
behaviour is tied to the rotation mode and how it handles the actual
placement. Will need to investigate further as well as ask LeonBlade
for assistance possibly.